corrected case^ error in calculator.c And added useful comments

+++ divide.c
@@ -3,9 +3,9 @@
 #include<stdio.h>
 int main()
 {
-  int n1,n2,div;
+  float n1,n2,div;
   printf("Enter two numbers:");// Take input from user
-  scanf("%d%d",&n1,&n2);
+  scanf("%f%f",&n1,&n2);
   div=n1/n2; // for divide users number
-  printf("Division of two numbers=%d",div);// give output of the users input 
+  printf("Division of two numbers=%.3f",div);// give output of the users input
 }
