+++ divisionalt.c
@@ -1,9 +1,9 @@
 #include<stdio.h>
 int main()
 {
-  int a,b,div;   // variable declaration
+  float a,b,div;   // variable declaration
   printf("Enter values of two numbers:");
-  scanf("%d%d",&a,&b);
+  scanf("%f%f",&a,&b);
   if(b==0)
   {
     printf("Infinity");
@@ -11,6 +11,6 @@ int main()
   else
   {
     div=a/b;
-    printf("Division of two numbers=%d",div);
+    printf("Division of two numbers=%.3f",div);
   }
 }
